+++ src/main/java/org/pahappa/systems/hms/models/Patient.java
@@ -1,7 +1,7 @@
-package models;
+package org.pahappa.systems.hms.models;
 
 import jakarta.persistence.*;
-import constants.UserRole;
+import org.pahappa.systems.hms.constants.UserRole;
 import org.hibernate.annotations.Where;
 
 import java.util.ArrayList;
